% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_summary.R
\name{df_summary}
\alias{df_summary}
\title{Averages and SEs summary of fata frames}
\usage{
df_summary(df_list)
}
\arguments{
\item{df_list}{Data Frame or list of data frames. If data frame, it will be placed into a list.}
}
\value{
Returns a list of data frames, one DF for each of the ones in the input list. Made to work with output from \link{split_dataframe}.
}
\description{
Used in the replication and extension of Table 1 in Orr, Fowler and Huber (2023). Outputs summary data frames with averages and SEs grouped in the same way as in the paper.
}
\examples{
NULL

}
