% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diffs.R
\name{extract_diffs}
\alias{extract_diffs}
\title{Statistical significance of differences}
\usage{
extract_diffs(df_list)
}
\arguments{
\item{df_list}{Data frame or list of data frames. If data frame, it will be placed into a list and the output will be identical to that in the article by Orr, Fowler and Huber (2023).}
}
\value{
Returns a list of coefficients, SEs and p-values for each of the variables in the data frames. These are grouped based on whether the respondent is evaluating a copartisan or an out-partisan.
}
\description{
Takes a data frame or a list of data frames (created by \link{split_dataframe}), and returns a list of coefficients (difference between the averages), the SEs and p-values, used to assess statistical significance of the differences between groups.
}
