% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.summ.R
\name{find.summ}
\alias{find.summ}
\title{Search table with SEs and averages}
\usage{
find.summ(summ, var, match_party, policy_type, match_policy)
}
\arguments{
\item{summ}{Summary table to search}

\item{var}{Variable of affect to search for}

\item{match_party}{Value of 'match_party' to search for.}

\item{policy_type}{Value of 'policy_type' to search for.}

\item{match_policy}{Value of 'match_policy' to search for.}
}
\value{
Returns the value in the table
}
\description{
Searches the output of \link{df_summary} for a specific value. Used in \link{maketable1}. This function was only made to save space in the main script.
}
\examples{
NULL

}
