% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table1.R
\name{generate_table1}
\alias{generate_table1}
\title{Generate table 1 or extension}
\usage{
generate_table1(
  summaries,
  diffs,
  caption = NULL,
  label = NULL,
  pos = "tbp",
  split.var = NULL,
  custom.table.names = NULL,
  skip.table.names = FALSE,
  note = NULL
)
}
\arguments{
\item{summaries}{List. Summary table/s from \link{df_summary}}

\item{diffs}{List. Differences in means, SEs and p-values from \link{extract_diffs}}

\item{caption}{Character. Caption for the table}

\item{label}{Character. Label for the table}

\item{pos}{Character. Position of the table}

\item{split.var}{Character. Variable to split the table by}

\item{custom.table.names}{Vector. Custom names for the tables}

\item{skip.table.names}{Logical. Skip the table names}

\item{note}{Character. Note to add to the table}
}
\value{
Returns a LaTeX table
}
\description{
Generates a LaTeX table based on Table 1 in Orr, Fowler and Huber (2023) or an extension of it. The table is based on the output from \link{df_summary} and \link{extract_diffs}. This function is a part of the \link{maketable1} function.
}
