% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketable1.R
\name{maketable1}
\alias{maketable1}
\title{Replicate Table 1 from the paper}
\usage{
maketable1(
  df,
  caption = NULL,
  label = NULL,
  pos = "tbp",
  split.var = NULL,
  custom.table.names = NULL,
  skip.table.names = FALSE,
  note = NULL
)
}
\arguments{
\item{df}{Data frame.}

\item{caption}{Character. Caption for the table.}

\item{label}{Character. Label for the table.}

\item{pos}{Character. Position of the table.}

\item{split.var}{Variable that the dataframe should be split on before generating the table. If NULL, the table will be generated for the entire dataset.}

\item{custom.table.names}{Vector. Custom names for the tables.}

\item{skip.table.names}{Logical. Skip the table names.}

\item{note}{Character. Note to add to the table.}
}
\value{
Returns a LaTeX table.
}
\description{
Requires the booktabs LaTeX package. Allows both for identical replication, and extension of Table 1 from Orr, Fowler and Huber (2023). The function can be used to generate the table for the entire dataset, or for a split dataset based on a specified variable. This function combines \link{df_summary}, \link{extract_diffs}, and \link{generate_table1} and automatically ties them together.
}
