% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_stars.R
\name{p_stars}
\alias{p_stars}
\title{Get number of stars OR thresholds for stars}
\usage{
p_stars(p_val, stars = c(0.001, 0.01, 0.05))
}
\arguments{
\item{p_val}{String or numeric. If numeric, the function outputs the number of stars. If "legend", the output is a full legend for tables. If '\emph{\strong{", "}", or "}", the output is the threshold.
, the function outputs the threshold for the number of stars.}

\item{stars}{Vector. Sets the thresholds for the stars. Default is 0.001 for 3 stars, 0.01 for 2, 0.05 for 1.}
}
\value{
Returns either full legend for tables, number of stars, or a specific threshold for a number of stars.
}
\description{
Used in \link{maketable1} and \link{makedifftables} to construct the legends.
}
\examples{
p_stars(0.001)
p_stars("**")
p_stars("legend")

}
