% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataframe.R
\name{split_dataframe}
\alias{split_dataframe}
\title{Split a DF into two based on a specified variable}
\usage{
split_dataframe(df, split_var)
}
\arguments{
\item{df}{Data frame to split}

\item{split_var}{String. Name of dummy variable that df should be split on}
}
\value{
Returns a list of two data frames + third item with the name of the split variable
}
\description{
Used in calculations for \link{maketable1} and \link{makedifftables}. This step is not necessary when replicating the tables, but it is used for the extension to compare two different groups.
}
\examples{
NULL

}
